Port ckanext-aquacross_fip_waf_harvest to this plugin
The ckanext-aquacross_fip_waf_harvest plugin copied and override whole
modules of the ckanext-spatial extension in order to modify the
harvested dataset dicts. That means there was a lot of duplicated code
that does not receive upstream fixes.
The recommended way of doing this is implementing the
`ISpatialHarvester` interface, which hooks into the harvesting process
to provide the harvested values and the generated data dict so it can be
modified by other plugins, which is exactly what we need to do.